Allow resetting local author to original one
After a pair programming session, the author should be able to go back
and redefine local author name to be only him again:

    $ git pair ended

If no global git configuration is used and an author cannot be
identified an error will be thrown asking the user to configure Git
properly.

Usage message given when no argument is given to the script now
displayes proper explanation on which arguments are accepted and what
they are.

Every `git pair` command now displays a feedback on what current authors
are configured so people using the script know the results before a
commit message is created.