Tests current implementation with Bats
Use tests to formalize current behavior, even if it is not what one
would expected or like to have. These behaviors are made explicit on
tests as well as with @TODO annotations on top of them.

Along with the different test suites (*.bats), there are helper scripts
which help me (and hopefully others) to relate better to these tests:

- `assertions.bash` provide xunit-like assertion style functions.
- `gitconfig` is used as the "global Git user" configuration placed on
  the user's $HOME directory.
- `git_environment.bash` overwrites $HOME variable to Bats temporary
  directory and moves git configuration there. It also creates a dummy
  repository where we can execute our suite into.

Travis-CI support is made available as well as the infamous badge of the
current build status of it. One can always see how to install Bats and
execute the test suite from `travis.yml` file.