Sorts author names instead of just appending them
By keeping pair names sorted, whenever you pair with the same people you
should get the same author. Prior, the first author in the list was
always the author's machine which could make the same pair yield
different author name results depending on the machine being used to
produce the commits.

The lines to sort authors are a mess, that is why I was delaying this.
As a remind to myself, this is what they do:

* Echoes final author name with all pairs.
* Uses `tr` to change "+" into new lines.
* Uses `sed` to remove leading white spaces from author names.
* Uses `sed` to remove trailing white spaces from author names.
* Sorts all author names available.
* Uses `tr` to change new lines back to "+".
* Uses `sed` to remove trailing "+" form end of the line.
* Uses `sed` to add spaces before and after "+".

I am not proud of it but I made sure that this works on Linuxes and BSDs
(OSX).