Adds one line install using Sinister
Although Sinister asks for a sudo without needing it, it is just a
matter of sending a pull request to it. All in all, it is way easier to
install using it for people who do not know their way into the shell
that much.

Short URL `https://git.io/git-pair` will always point to the master
`git-pair` file. I guess there is no renaming it or I am hoping symbolic
links will work if necessary - thinking about the file name, I hate how I
missed the .sh extension to it. Oh my, lets hope that ship has not
sailed yet.